finish update of indexing to size_t

+++ include/grampm_integrators.hpp
@@ -7,12 +7,12 @@
 namespace GraMPM {
     namespace integrators {
         template<typename F>
-        void MUSL(GraMPM::MPM_system<F> &p, const F &dt, const int &max_timestep, 
-            const int &print_timestep_interval, const int &save_timestep_interval) {
+        void MUSL(GraMPM::MPM_system<F> &p, const F &dt, const size_t &max_timestep, 
+            const size_t &print_timestep_interval, const size_t &save_timestep_interval) {
 
             #pragma omp parallel default(shared)
             {
-            for (int itimestep = 1; itimestep < max_timestep+1; ++itimestep) {
+            for (size_t itimestep = 1; itimestep < max_timestep+1; ++itimestep) {
 
                 // print to terminal
                 if (itimestep % print_timestep_interval == 0) {
